+++ changelog.md
@@ -1,5 +1,11 @@
-# 1.0.0
+# 1.0.0-beta.2
+- Support for GPU acceleration
+- Adjustable `Detection Threshold`
+- Option to mirror the video
+- Optimizations
+
+# 1.0.0-beta.1
 - Camera inside the game
 - Pose estimation
 - Control the game with your pose
-- 
+